Revert earlier color styling changes for now
The panels should appear to be below the main screen, not above it, plus
there are tons of other styling tweaks on master already so I'm reverting
most of these except for basic ones rather than make it a random mix of
two sets of tweaks. If I accidentally threw out any changes that are still
relevant then please recommit them.